// src/types/generated/ModelParameterOverrides.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

/**
 * Model parameter overrides for a conversation
 */
export type ModelParameterOverrides = { temperature?: number, max_tokens?: number, top_p?: number, frequency_penalty?: number, presence_penalty?: number, };